move controller to specified file
